403 for internal apis; moved users dtos in the correct package; gets do not mutate data
