+++ backend/export-sv/api/src/main/kotlin/com/r8n/backend/export/api/dto/UserCompleteDataDto.kt
@@ -4,8 +4,8 @@ import com.r8n.backend.core.api.PageResponseDto
 import com.r8n.backend.mock.api.dto.SupportThreadDto
 import com.r8n.backend.opinions.api.access.dto.AccessRequestDto
 import com.r8n.backend.opinions.api.lists.dto.OpinionListDto
-import com.r8n.backend.users.api.dto.ConsentDto
-import com.r8n.backend.users.api.dto.PersonalIdentifiableInformationSectionDto
+import com.r8n.backend.users.integration.api.dto.ConsentDto
+import com.r8n.backend.users.integration.api.dto.PersonalIdentifiableInformationSectionDto
 import com.r8n.backend.users.api.dto.UserStatusEnumDto
 import java.time.Instant
 import java.util.UUID
