+++ backend/export-sv/service/src/main/kotlin/com/r8n/backend/export/service/DataExportService.kt
@@ -7,7 +7,7 @@ import com.r8n.backend.mock.api.MessagingApi
 import com.r8n.backend.opinions.api.access.IncomingAccessRequestApi
 import com.r8n.backend.opinions.api.access.OutgoingAccessRequestApi
 import com.r8n.backend.opinions.integration.api.OpinionListsInternalApi
-import com.r8n.backend.users.api.dto.PersonalIdentifiableInformationSectionDto
+import com.r8n.backend.users.integration.api.dto.PersonalIdentifiableInformationSectionDto
 import com.r8n.backend.users.integration.api.UsersInternalApi
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
