+++ backend/gateway-sv/service/src/main/resources/application.yml
@@ -13,6 +13,12 @@ spring:
       server:
         webflux:
           routes:
+            - id: block_internal
+              uri: no-forward://internal
+              predicates:
+                - Path=/api/internal/**
+              filters:
+                - SetStatus=403
             - id: auth
               uri: ${INTERSERVICE_PROTOCOL}://${SERVICES_USERS_HOST}:${SERVICES_USERS_PORT}
               predicates:
