+++ backend/gateway-sv/service/src/test/resources/application-test-e2e.yml
@@ -13,6 +13,10 @@ spring:
       server:
         webflux:
           routes:
+            - id: block_internal
+              uri: no-forward://internal
+              predicates:
+                - Path=/api/internal/**
             - id: auth
               uri: ${INTERSERVICE_PROTOCOL}://${SERVICES_USERS_HOST}:${SERVICES_USERS_PORT}
               predicates:
