+++ backend/opinions-sv/api-integration/src/main/kotlin/com/r8n/backend/opinions/integration/api/OpinionListsInternalApi.kt
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping
 
 interface OpinionListsInternalApi {
     companion object {
-        const val MINE_FULL_PATH = "/api/opinion-lists/mine/full"
+        const val MINE_FULL_PATH = "/api/internal/opinion-lists/mine/full"
     }
 
     @GetMapping(MINE_FULL_PATH)
