+++ backend/opinions-sv/api/src/main/kotlin/com/r8n/backend/opinions/api/access/OutgoingAccessRequestApi.kt
@@ -6,6 +6,7 @@ import com.r8n.backend.opinions.api.access.dto.AccessRequestDto
 import com.r8n.backend.opinions.api.access.dto.RequestStatusEnumDto
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestParam
 import java.time.Instant
 import java.util.UUID
@@ -29,12 +30,12 @@ interface OutgoingAccessRequestApi {
         pageable: PageRequestDto,
     ): PageResponseDto<AccessRequestDto>
 
-    @GetMapping(CREATE_PATH)
+    @PostMapping(CREATE_PATH)
     fun create(
         @PathVariable listId: UUID,
     ): AccessRequestDto
 
-    @GetMapping(CANCEL_PATH)
+    @PostMapping(CANCEL_PATH)
     fun cancel(
         @PathVariable requestId: UUID,
     ): AccessRequestDto
