+++ backend/users-sv/api-integration/src/main/kotlin/com/r8n/backend/users/integration/api/dto/UserDto.kt
@@ -1,5 +1,6 @@
-package com.r8n.backend.users.api.dto
+package com.r8n.backend.users.integration.api.dto
 
+import com.r8n.backend.users.api.dto.UserStatusEnumDto
 import java.time.Instant
 import java.util.UUID
