+++ backend/users-sv/client/src/main/kotlin/com/r8n/backend/users/integration/UsersRestClient.kt
@@ -2,8 +2,8 @@ package com.r8n.backend.users.integration
 
 import com.r8n.backend.core.api.PageRequestDto
 import com.r8n.backend.core.api.PageResponseDto
-import com.r8n.backend.users.api.dto.UserDto
-import com.r8n.backend.users.api.dto.UserSessionDto
+import com.r8n.backend.users.integration.api.dto.UserDto
+import com.r8n.backend.users.integration.api.dto.UserSessionDto
 import com.r8n.backend.users.integration.api.UsersInternalApi
 import com.r8n.backend.users.integration.api.UsersInternalApi.Companion.ID_PATH
 import com.r8n.backend.users.integration.api.UsersInternalApi.Companion.IS_ADMIN_PATH
